+++ theano/gof/cc.py
@@ -11,8 +11,8 @@ from io import StringIO
 
 import numpy as np
 
-from theano import config
-from theano.gof import cmodule, graph, link, utils
+from theano import config, link
+from theano.gof import cmodule, graph, utils
 from theano.gof.callcache import CallCache
 from theano.gof.compilelock import get_lock, release_lock
